backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.5
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.7
chromadb==0.4.24
faiss-cpu==1.8.0
numpy>=1.26.0
docling>=1.7.0
typing-extensions==4.8.0
